requirements.txt
streamlit>=1.28.0
SpeechRecognition>=3.10.0
google-generativeai>=0.3.0
gtts>=2.3.2